Move the rotation functions to the rotations file
