+++ pulse/utils/common_utils.py
@@ -200,251 +200,6 @@ def inverse_matrix_3x3(A):
     
     return invA
 
-def transformation_matrix_3x3(delta_x, delta_y, delta_z, gamma=0):
-    '''    
-    This method returns the rotation matrix of an element 
-    based on its spatial position. 
-    
-    Parameters
-    ----------
-    delta_x: int, float
-        value in meters
-    
-    delta_y: int, float
-        value in meters
-
-    delta_z: int, float
-        value in meters
-
-    Returns
-    -------
-    out: numpy.ndarray(3,3)
-        rotation matrix
-
-    '''
-
-    L_ = np.sqrt(delta_x**2 + delta_y**2)
-    L  = np.sqrt(delta_x**2 + delta_y**2 + delta_z**2)
-
-    cossine_epsilon = L_ / L
-    sine_epsilon = - delta_z / L
-    
-    if L_ > 0.0001*L:
-        sine_delta = delta_y/L_
-        cossine_delta = delta_x/L_
-    else:
-        sine_delta = 0
-        cossine_delta = 1
-    
-    cossine_gamma = np.cos(gamma)
-    sine_gamma = np.sin(gamma)
-
-    # Matrices product order - Rx@Ry@Rz (@Palazzolo, A. Vibration theory and applications with finite element and active vibration control. pg 677)
-    rotation_matrix = np.array([    [   cossine_delta * cossine_epsilon, 
-                                       sine_delta * cossine_epsilon, 
-                                        -sine_epsilon   ], 
-                                    [   cossine_delta * sine_epsilon * sine_gamma - sine_delta * cossine_gamma,
-                                        sine_delta * sine_epsilon * sine_gamma + cossine_delta * cossine_gamma,
-                                        cossine_epsilon * sine_gamma    ],
-                                    [   cossine_delta * sine_epsilon * cossine_gamma + sine_delta * sine_gamma,
-                                        sine_delta * sine_epsilon * cossine_gamma - cossine_delta * sine_gamma,
-                                        cossine_epsilon * cossine_gamma ]    ]) 
-
-    return rotation_matrix
-
-
-def transformation_matrix_3x3xN(delta_x, delta_y, delta_z, gamma=0):
-    '''    
-    This method returns the rotation matrices to a set of N elements 
-    based on their spatial positions. 
-    
-    Parameters
-    ----------
-    delta_x: numpy.ndarray
-        values in meters
-    
-    delta_y: numpy.ndarray
-        values in meters
-
-    delta_z: numpy.ndarray
-        values in meters
-
-    Returns
-    -------
-    out: numpy.ndarray(N,3,3)
-        rotation matrix
-
-    '''
-
-    number_elements = len(delta_x)
-    L_ = np.sqrt(delta_x**2 + delta_y**2)
-    L  = np.sqrt(delta_x**2 + delta_y**2 + delta_z**2)
-    
-    cossine_gamma = np.cos(gamma)
-    sine_gamma = np.sin(gamma)
-
-    sine_delta = np.zeros(number_elements, dtype=float)
-    cossine_delta = np.zeros(number_elements, dtype=float)
-
-    for i in range(number_elements):
-
-        if L_[i] > 0.0001*L[i]:
-            sine_delta[i] = delta_y[i]/L_[i]
-            cossine_delta[i] = delta_x[i]/L_[i]
-        else:
-            sine_delta[i] = 0
-            cossine_delta[i] = 1
-
-    cossine_epsilon = L_ / L
-    sine_epsilon = - delta_z / L
-    
-    # Matrices product order - Rx@Ry@Rz (@Palazzolo, A. Vibration theory and applications with finite element and active vibration control. pg 677)
-    data_rot = np.array([   cossine_delta * cossine_epsilon, 
-                            sine_delta * cossine_epsilon, 
-                            -sine_epsilon, 
-                            cossine_delta * sine_epsilon * sine_gamma - sine_delta * cossine_gamma,
-                            sine_delta * sine_epsilon * sine_gamma + cossine_delta * cossine_gamma,
-                            cossine_epsilon * sine_gamma,
-                            cossine_delta * sine_epsilon * cossine_gamma + sine_delta * sine_gamma,
-                            sine_delta * sine_epsilon * cossine_gamma - cossine_delta * sine_gamma,
-                            cossine_epsilon * cossine_gamma   ])
-
-    return data_rot.T.reshape(-1,3,3)
-
-def transformation_matrix_3x3_by_angles(gamma, epsilon, delta):
-    '''    
-    This method returns the rotation matrix of an element based on 
-    the angles of rotations gamma, epsilon and delta. 
-    
-    Parameters
-    ----------
-    gamma: int, float
-        values in radians
-    
-    epsilon: int, float
-        values in radians
-
-    delta: int, float
-        values in radians
-
-    Returns
-    -------
-    out: numpy.ndarray(3,3)
-        rotation matrix
-
-    '''
-
-    sine_delta = np.sin(delta)
-    cossine_delta = np.cos(delta)
-
-    sine_epsilon = np.sin(epsilon)
-    cossine_epsilon = np.cos(epsilon)
-
-    sine_gamma = np.sin(gamma)
-    cossine_gamma = np.cos(gamma)
-
-    # Matrices product order - Rx@Ry@Rz (@Palazzolo, A. Vibration theory and applications with finite element and active vibration control. pg 677)
-    data_rot = np.array([   cossine_delta * cossine_epsilon, 
-                            sine_delta * cossine_epsilon, 
-                            -sine_epsilon, 
-                            cossine_delta * sine_epsilon * sine_gamma - sine_delta * cossine_gamma,
-                            sine_delta * sine_epsilon * sine_gamma + cossine_delta * cossine_gamma,
-                            cossine_epsilon * sine_gamma,
-                            cossine_delta * sine_epsilon * cossine_gamma + sine_delta * sine_gamma,
-                            sine_delta * sine_epsilon * cossine_gamma - cossine_delta * sine_gamma,
-                            cossine_epsilon * cossine_gamma   ])
-
-    return data_rot.reshape(3,3)
-
-def transformation_matrix_Nx3x3_by_angles(gamma, epsilon, delta):
-    '''    
-    This method returns the rotation matrices to a set of N elements 
-    based on the angles of rotations gamma, epsilon and delta. 
-    
-    Parameters
-    ----------
-    gamma: numpy.ndarray
-        values in radians
-    
-    epsilon: numpy.ndarray
-        values in radians
-
-    delta: numpy.ndarray
-        values in radians
-
-    Returns
-    -------
-    out: numpy.ndarray(N,3,3)
-        rotation matrix
-
-    '''
-
-    sine_delta = np.sin(delta)
-    cossine_delta = np.cos(delta)
-
-    sine_epsilon = np.sin(epsilon)
-    cossine_epsilon = np.cos(epsilon)
-
-    sine_gamma = np.sin(gamma)
-    cossine_gamma = np.cos(gamma)
-
-    # Matrices product order - Rx@Ry@Rz (@Palazzolo, A. Vibration theory and applications with finite element and active vibration control. pg 677)
-    data_rot = np.array([   cossine_delta * cossine_epsilon, 
-                            sine_delta * cossine_epsilon, 
-                            -sine_epsilon, 
-                            cossine_delta * sine_epsilon * sine_gamma - sine_delta * cossine_gamma,
-                            sine_delta * sine_epsilon * sine_gamma + cossine_delta * cossine_gamma,
-                            cossine_epsilon * sine_gamma,
-                            cossine_delta * sine_epsilon * cossine_gamma + sine_delta * sine_gamma,
-                            sine_delta * sine_epsilon * cossine_gamma - cossine_delta * sine_gamma,
-                            cossine_epsilon * cossine_gamma   ])
-
-    return data_rot.T.reshape(-1,3,3)
-
-
-# def error( msg, title = " ERROR "):
-#     '''
-#     PyQt5 error message.
-
-#     Parameters
-#     ----------
-#     msg: str
-#         text to be displayed.
-
-#     title: str
-#         window title.
-#     '''
-
-#     msg_box = QMessageBox()
-#     msg_box.setWindowFlags(Qt.WindowStaysOnTopHint)
-#     # msg_box.setWindowModality(Qt.WindowModal)
-#     msg_box.setIcon(QMessageBox.Critical)
-#     msg_box.setText(msg)
-#     msg_box.setWindowTitle(title)
-#     msg_box.exec_()
-
-
-# def info_messages(msg, title = " INFORMATION "):
-#     '''
-#     PyQt5 info message.
-
-#     Parameters
-#     ----------
-#     msg: str
-#         text to be displayed.
-
-#     title: str
-#         window title.
-#     '''
-
-#     msg_box = QMessageBox()
-#     msg_box.setWindowFlags(Qt.WindowStaysOnTopHint)
-#     # msg_box.setWindowModality(Qt.WindowModal)
-#     msg_box.setIcon(QMessageBox.Information)
-#     msg_box.setText(msg)
-#     msg_box.setWindowTitle(title)
-#     msg_box.exec_()
-
 def getColorRGB(color):
     temp = color[1:-1] #Remove "[ ]"
     tokens = temp.split(',')
